Organize directories
Rename sfml-sorting subdirectory to src.

Moved libraries to their own separate subdirectories within src.
