+++ CMakeLists.txt
@@ -15,4 +15,4 @@ set(CMAKE_LIBRARY_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 set(CMAKE_ARCHIVE_OUTPUT_DIRECTORY ${CMAKE_BINARY_DIR}/lib)
 
 # Include sub-projects.
-add_subdirectory ("sfml-sorting")
+add_subdirectory ("src")
