+++ src/utils/CMakeLists.txt
@@ -1,0 +1,3 @@
+﻿cmake_minimum_required (VERSION 3.8)
+
+add_library(utils RandomNumberGenerator.cpp)
